refactor(468): replace useWishlistProductIds with useWishlistProducts hook for improved wishlist management
